// src/app/model/reporte.model.ts
export class Reporte {
  numero: string;
  descTipoDoc: string;
  numDocumentoPersona: string;
  nombres: string;
  txtCorreoElectronico: string;
  txtDescripcionTramite: string;
  descEstaTrami: string;
  fecRegistroTramite: string;
  fecEnvioTramite: string;
  fecEnviaDesd: string;
  fecEnviaHast: string;
  fecRegistroDesd: string;
  fecRegistroHast: string;
  codEtipoDocumento: string;
  txtRazonSocial: string;
  codEestaTrami: string;
}